Revert "微修正"
This reverts commit 0a268b2f1d576b10920ae3754bdcd8876d456e90.

+++ model/items_test.go
@@ -59,7 +59,7 @@ func TestRegisterOwner(t *testing.T) {
 		owner.Count = 5
 		item, err := RegisterOwner(owner, item)
 
-		assert.Equal(uint(6), item.Owners[0].Count)
+		assert.Equal(6, item.Owners[0].Count)
 
 		assert.NoError(err)
 		assert.NotEmpty(item)
